Extract levelshots directly to the ouptut directory
It is annoying to manage a separate area, and we need two paths: one
relative to the working directory for the zip extractor to write the
file, and one relative to the output directory for the html
hyperreference. To complicate things further the zip extractions retains
the inner directory structure - where there is always a "levelshots"
directory.